ENH: Move thredds connection to a new module
