+++ tabs/app.py
@@ -1,12 +1,12 @@
 import argparse
 import sys
-from threading import Timer, RLock
 
 import numpy as np
 from flask import Flask, jsonify, make_response, redirect, request, url_for
 from flask.ext.compress import Compress
 
 from tabs import thredds_frame_source
+from tabs.thredds_connection import ThreddsConnection
 
 
 class ReverseProxied(object):
@@ -54,72 +54,9 @@ DECIMATE_FACTOR = 10
 RANDOM_STATE = np.random.get_state()
 
 
-# We should probably maintain a connection for at least a short while
-class THREDDS_CONNECTION(object):
-
-    def __init__(self, timeout=300.0, **fs_args):
-        """ Create an expiring connection to the THREDDS server.
-
-        The connection will drop after 5 minutes of non-use. Any subsequent
-        attempt to use the connection will initiate a new one. Access to the
-        connection is RLock'd to ensure only one connection is alive at a time.
-
-        Parameters:
-        timeout : int, seconds
-            The lenght of time in seconds to hold open a connection.
-
-        Remaining keyword args are passed to the connection's constructor.
-        """
-        self._fs = None
-        self._fs_lock = RLock()
-        self._fs_args = fs_args
-        self._timer = None
-        self.timeout = float(timeout)
-
-    def _forget(self):
-        app.logger.info("Closing THREDDS connection")
-        if self._timer:
-            self._timer.cancel()
-            self._timer = None
-        self._fs = None
-
-    def _reset_timer(self):
-        app.logger.info("Resetting THREDDS connection timer")
-        if self._timer:
-            self._timer.cancel()
-        self._timer = Timer(self.timeout, self._forget)
-        self._timer.start()
-
-    def fs():
-        doc = "The fs property."
-
-        def fget(self):
-            with self._fs_lock:
-                if not self._fs:
-                    app.logger.info("Opening new THREDDS connection")
-                    # Ensure that we get the same ordering of grid points
-                    np.random.set_state(RANDOM_STATE)
-                    cls = thredds_frame_source.THREDDSFrameSource
-                    self._fs = cls(**self._fs_args)
-                    app.logger.info("THREDDS connection ready")
-                self._reset_timer()
-                return self._fs
-
-        def fset(self, value):
-            with self._fs_lock:
-                self._fs = value
-                self._reset_timer()
-                return self._fs
-
-        def fdel(self):
-            with self._fs_lock:
-                self._forget()
-        return locals()
-    fs = property(**fs())
-
-
-tc = THREDDS_CONNECTION(data_uri=thredds_frame_source.DEFAULT_DATA_URI,
-                        decimate_factor=DECIMATE_FACTOR)
+tc = ThreddsConnection(app, random_state=RANDOM_STATE,
+                       data_uri=thredds_frame_source.DEFAULT_DATA_URI,
+                       decimate_factor=DECIMATE_FACTOR)
 
 
 def jsonify_dict_of_array(obj):
